requirements.txt
aenum==3.1.0
asgiref==3.4.1
Django==3.2.7
django-cors-headers==3.8.0
django-js-asset==1.2.2
django-mdeditor==0.1.18
django-ratelimit==3.0.1
django-redis==5.0.0
django-simpleui==2021.8.17
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
Markdown==3.3.4
mysql-client==0.0.1
mysqlclient==2.0.3
Pillow==8.3.2
PyJWT==2.3.0
pytz==2021.1
redis==3.5.3
sqlparse==0.4.2
